app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Video LLM-based search engine"
authors = ["MichalPysik <n3r0x541@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
pillow = "^11.0.0"
torch = "^2.5.1"
torchvision = "^0.20.1"
pydantic = "^2.10.3"
python-dotenv = "^1.0.1"
opencv-python = "^4.10.0.84"
pymilvus = "^2.5.0"
poetry-dotenv-plugin = "^0.2.0"
sentencepiece = "^0.2.0"
transformers = "^4.47.0" # comment out this line and the following line (timm) when using Blip model
timm = "1.0.13"
# salesforce-lavis = "^1.0.2" # comment out this line and the following line (tokenizers) when using CLIP, SigLIP or ALIGN model
# tokenizers = "0.13.2" # requires RUST compiler 1.72.0!!! (RUSTUP_TOOLCHAIN=1.72.0 poetry add tokenizers==0.13.2)
minio = "^7.2.15"
ffmpeg-python = "^0.2.0"
bitsandbytes = "^0.45.3"
accelerate = "^1.4.0"
openai = "^1.65.2"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
